// packages/pf2etools-data/src/schemas/lib/helpers.zod.ts
export const zTypedNumberSchema = z.strictObject({
	unit: z.string().optional(),
	customUnit: z.string().optional(),
	number: z.number().or(z.string()).optional(),
	entry: z.string().optional(),
});
// ...
export const zDurationSchema = z.union([
	z.strictObject({
		unit: z.string().optional(),
		customUnit: z.string().optional(),
		number: z.number().or(z.string()).optional(),
		entry: z.string().optional(),
		dismiss: z.boolean().optional(),
		sustained: z.boolean().optional(),
	}),
	z.strictObject({
		type: z.string(),
		entry: z.string(),
		duration: zTypedNumberSchema,
	}),
]);

export type OtherSource = z.infer<typeof zOtherSourceSchema>;
export const zOtherSourceSchema = z.strictObject({
	Expanded: z.string().array().optional(),
	Reprinted: z.string().array().optional(),
	Originally: z.string().array().optional(),
});